decoupled UI from processing
