+++ C-BPSK Demodulator/src/constellation.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include <wx/wxprec.h>
 #ifndef WX_PRECOMP
 #include <wx/wx.h>
