+++ C-BPSK Demodulator/src/dsp/noaa_deframer.h
@@ -1,3 +1,4 @@
+#pragma once
 #include <cstdint>
 #include <vector>
