+++ C-BPSK Demodulator/src/window.h
@@ -1,5 +1,5 @@
 #pragma once
-
+#include <functional>
 #include <wx/wxprec.h>
 #ifndef WX_PRECOMP
 #include <wx/wx.h>
@@ -17,6 +17,7 @@
 #include <dsp/moving_average.h>
 #include <dsp/pipe.h>
 #include <dsp/noise_source.h>
+#include "dsp.h"
 #include "dsp/noaa_deframer.h"
 
 #define BUFFER_SIZE 4096
@@ -44,52 +45,12 @@ public:
     wxButton *presetNOAAHRPT, *presetMeteorHRPT;
 
 private:
+    CBPSKDemodulatorDsp *dsp;
     std::string inputFilePath, outputFilePath;
-    void updateProgress(int current, int total);
+    void updateProgress(int current, int total, int frame_count);
     void done();
-
-private:
-    // All buffers we use along the way
-    std::complex<float> *buffer;
-    std::complex<float> *agc_buffer;
-    std::complex<float> *front_rrc_buffer;
-    std::complex<float> *front_rrc_buffer2;
-    std::complex<float> *pll_buffer;
-    float *pll_buffer2;
-    float *rrc_buffer;
-    float *rrc_buffer2;
-    float *recovery_buffer;
-    float *recovery_buffer2;
-    float *recovered_buffer;
-    float *noise_buffer;
-    uint8_t *bitsBuffer;
-    // Int16 buffer
-    int16_t *buffer_i16;
-    // Int8 buffer
-    int8_t *buffer_i8;
-    // Uint8 buffer
-    uint8_t *buffer_u8;
-
-private:
-    // All blocks used along the way
-    libdsp::AgcCC *agc;
-    libdsp::FIRFilterCCF *front_rrc_fir_filter;
-    libdsp::BPSKCarrierPLL *carrier_pll;
-    libdsp::FIRFilterFFF *rrc_fir_filter;
-    libdsp::MovingAverageFF *movingAverage;
-    libdsp::ClockRecoveryMMFF *clock_recovery;
-    libdsp::NoiseSource *noise_source;
-    NOAADeframer *noaa_deframer_blk;
-    
-
-private:
-    // All FIFOs we use along the way
-    libdsp::Pipe<std::complex<float>> *file_pipe;
-    libdsp::Pipe<std::complex<float>> *agc_pipe;
-    libdsp::Pipe<std::complex<float>> *front_rrc_pipe;
-    libdsp::Pipe<float> *pll_pipe;
-    libdsp::Pipe<float> *rrc_pipe;
-    libdsp::Pipe<float> *recovery_pipe;
+    void updateConstellation(int8_t a, int8_t b, int i);
+    bool noaaMode, meteorMode;
 };
 
 class CBPSKDemodulatorFrame : public wxFrame
@@ -97,48 +58,3 @@ class CBPSKDemodulatorFrame : public wxFrame
 public:
     CBPSKDemodulatorFrame();
 };
-
-class CBPSKDemodulatorDsp
-{
-public:
-    void initDSP(function progressCallback, function doneCallback);
-    void startDSP(std::string inputFilePath,
-                  std::string outputFilePath,
-                  bool optionF32,
-                  bool optionI16,
-                  bool optionI8,
-                  bool optionW8,
-                  float samplerate,
-                  float symbolrate,
-                  float rrc_alpha,
-                  float rrc_taps,
-                  bool noaaMode,
-                  bool meteorMode);
-    void destroyDSP();
-
-private:
-    void fileThreadFunction();
-    void agcThreadFunction();
-    void frontRrcThreadFunction();
-    void pllThreadFunction();
-    void rrcThreadFunction();
-    void clockrecoveryThreadFunction();
-    void finalWriteThreadFunction();
-    void (*progressCallback)(int, int);
-    void (*doneCallback)();
-
-private:
-    std::thread *fileThread;
-    std::thread *agcThread;
-    std::thread *frontRrcThread;
-    std::thread *pllThread;
-    std::thread *rrcThread;
-    std::thread *clockrecoveryThread;
-    std::thread *finalWriteThread;
-    bool f32, i16, i8, w8;
-    float samplerate, symbolrate, rrc_alpha, rrc_taps;
-    size_t data_in_filesize;
-    std::ifstream data_in;
-    std::ofstream data_out;
-    bool noaa_deframer, rrc_filter;
-}
